refactor(validation): audit and strengthen the validation layer
Aligns validators with the four declared requirements and drops the
fragile label-sanitization fallback in favour of explicit @jpipe_link
bindings.

Changes:
- Req C: @jpipe now raises ValueError at decoration time when a function
  declares produce=[...] but does not have 'produce' as its last
  parameter.
- Req D: declared consumed variables that are never referenced in the
  function body (Load context) are now a ValueError at decoration time
  instead of a silent warning; AST visitor restricted to Load context
  to avoid false-positives from assignment targets.
- Req A/B: EvidenceDependencyValidator rewritten to iterate over graph
  nodes directly, covering sub-conclusion predecessors (skipped when
  unbound) and including element ID + label in every error message.
- Req E: new UnboundElementValidator rejects any evidence/strategy node
  that lacks an explicit @jpipe_link binding.
- validators.py split into a validators/ package (one file per
  validator) for readability; public import surface unchanged.
- All 8 e2e resource scenarios updated with @jpipe_link; two resources
  had latent bugs (self_dependency param mismatch, missing_consumer
  unused variable) that are fixed as a side-effect.

183/183 tests pass (123 unit + 60 e2e).
