+++ src/jpipe_runner/framework/decorators/jpipe_decorator.py
@@ -21,6 +21,7 @@ def jpipe(consume: Optional[List[str]] = None, produce: Optional[List[str]] = No
     produce = produce or []
 
     def decorator(func: Callable) -> Callable:
+        _check_produce_param(func, produce)
         consume_checker = _init_checker(ConsumedVariableChecker, func, consume)
         produce_checker = _init_checker(ProducedVariableChecker, func, produce)
 
@@ -41,6 +42,20 @@ def jpipe(consume: Optional[List[str]] = None, produce: Optional[List[str]] = No
         return wrapper
 
     return decorator
+
+
+def _check_produce_param(func: Callable, produce: List[str]) -> None:
+    if not produce:
+        return
+    params = list(inspect.signature(func).parameters)
+    if not params or params[-1] != "produce":
+        raise ValueError(
+            f"[jpipe] Function '{func.__name__}' declares produce={produce!r} "
+            f"but its last parameter is not 'produce'.\n"
+            f"  • Actual signature: {func.__name__}({', '.join(params)})\n"
+            f"  • Fix: add 'produce' as the final parameter, "
+            f"typed as Callable[[str, Any], None]."
+        )
 
 
 def _init_checker(
@@ -93,6 +108,15 @@ class ConsumedVariableChecker:
                 ctx._set(self.func_name, param, None, RuntimeContext.CONSUME)
                 GLOBAL_LOGGER.debug(f"[{self.func_name}] Variable '{param}' registered as CONSUME")
 
+        unused = set(self.declared_params) - self.used_params
+        if unused:
+            raise ValueError(
+                f"[jpipe] Function '{self.func_name}' declares consumed variable(s) "
+                f"{sorted(unused)!r} that are never referenced in the function body.\n"
+                f"  • Fix: either reference each variable in the function body, "
+                f"or remove it from consume=[]."
+            )
+
     def inject_arguments(self, kwargs: dict) -> dict:
         """
         Injects variable values from the context into the function’s keyword arguments.
@@ -102,10 +126,6 @@ class ConsumedVariableChecker:
         :raises ValueError: If a declared variable is missing from the context.
         """
         for param in self.declared_params:
-            if param not in self.used_params:
-                GLOBAL_LOGGER.warning(
-                    f"Consumed variable '{param}' is declared but not used in function '{self.func_name}'."
-                )
             value = ctx.get(param)
             GLOBAL_LOGGER.debug(
                 f"[{self.func_name}] Injecting consumed variable '{param}' = {value}"
@@ -136,7 +156,8 @@ class ConsumedVariableChecker:
                 self.used_vars: set[str] = set()
 
             def visit_Name(self, node: ast.Name) -> None:
-                self.used_vars.add(node.id)
+                if isinstance(node.ctx, ast.Load):
+                    self.used_vars.add(node.id)
 
         visitor = VarVisitor()
         visitor.visit(tree)
