+++ tests/e2e/resources/circular_dependency/circular.py
@@ -1,6 +1,8 @@
 from jpipe_runner.framework.decorators.jpipe_decorator import jpipe
+from jpipe_runner.framework.decorators.link_decorator import jpipe_link
 
 
+@jpipe_link("S1")
 @jpipe(consume=["var_b"], produce=["var_a"])
 def function_a(var_b: str, produce) -> bool:
     """Function A depends on B"""
@@ -8,6 +10,7 @@ def function_a(var_b: str, produce) -> bool:
     return True
 
 
+@jpipe_link("S2")
 @jpipe(consume=["var_a"], produce=["var_b"])
 def function_b(var_a: str, produce) -> bool:
     """Function B depends on A - creates cycle"""
@@ -18,4 +21,4 @@ def function_b(var_a: str, produce) -> bool:
 @jpipe(consume=["var_a", "var_b"])
 def final_check(var_a: str, var_b: str) -> bool:
     """Final validation"""
-    return True
+    return var_a and var_b
