+++ tests/e2e/resources/missing_consumer/math_operation.py
@@ -1,6 +1,8 @@
 from jpipe_runner.framework.decorators.jpipe_decorator import jpipe
+from jpipe_runner.framework.decorators.link_decorator import jpipe_link
 
 
+@jpipe_link("E1")
 @jpipe(produce=["number_a"])
 def generate_number_a(produce) -> bool:
     """Generate first number"""
@@ -8,6 +10,7 @@ def generate_number_a(produce) -> bool:
     return True
 
 
+@jpipe_link("E2")
 @jpipe(produce=["number_b"])
 def generate_number_b(produce) -> bool:
     """Generate second number"""
@@ -15,8 +18,9 @@ def generate_number_b(produce) -> bool:
     return True
 
 
-# Missing consume for number_a and number_b
+# Intentionally consumes a variable that has no producer
+@jpipe_link("S1")
 @jpipe(consume=["missing_var"])
 def add_numbers(missing_var: int) -> bool:
-    """Add two numbers - should fail due to missing consumer"""
-    return True
+    """Add two numbers - fails due to missing variable"""
+    return missing_var > 0
