+++ tests/e2e/resources/missing_producer/string_operation.py
@@ -1,12 +1,15 @@
 from jpipe_runner.framework.decorators.jpipe_decorator import jpipe
+from jpipe_runner.framework.decorators.link_decorator import jpipe_link
 
 
+# Intentionally calls produce() without declaring it in the decorator —
+# processed_string is never registered in ctx, so validate_string has no producer.
+@jpipe_link("S1")
 @jpipe(consume=["input_string"])
-def process_string(input_string: str, produce) -> bool:
+def process_string(input_string: str) -> bool:
     """Process a string"""
     processed = input_string.upper()
-    produce("processed_string", processed)
-    return True
+    return len(processed) > 0
 
 
 @jpipe(consume=["processed_string"])
